feat: parse "attribute" markers
Parse "attribute" markers in order to add a field to parent class
