+++ src/main/scala/fr/inria/hopla/ast/ASTField.scala
@@ -4,5 +4,5 @@ package fr.inria.hopla.ast
  * Representation of a class field in ASTClass
  * @author Jérémy Bossut, Jonathan Geoffroy
  */
-class ASTField(val name: String, fieldType: String) {
+class ASTField(val name: String, val fieldType: String) {
 }
