+++ src/test/resources/xsd/attribute.xsd
@@ -1,0 +1,7 @@
+<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns="http://www.w3.org/1999/xhtml" version="1.0" xml:lang="en" targetNamespace="http://www.w3.org/1999/xhtml" elementFormDefault="qualified">
+    <xs:element name="element">
+        <xs:complexType mixed="true">
+            <xs:attribute name="attribute" type="xs:string"/>
+        </xs:complexType>
+    </xs:element>
+</xs:schema>
